Polynomial orders option
Now the command allows 1, 4 or 8 arguments: E.g.: order(1), order(1 4 3 2) or order(1 4 3 2 1 2 3 4) are allowed (the last one only if IV-WAS specified), but order(1 2 3) is not allowed.
With 4 arguments, argument 1, 2, 3 and 4, is the order used to estimate E(Y_t-Y_{t-1}|D_{t-1}), P(S_{t}=0|D_{t-1}), P(S_{+, t}=1|D_{t-1}), and P(S_{-, t}=1|D_{t-1}), respectively.
With 8 arguments the same logic is applied, but the first 4 arguments are for the first stage, and the next 4 for the reduced form.
Finally, if IV-WASis requested but order has 4 arguments, we apply the same orders to first stage and reduced form.